+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project (aims to) adhere to [Semantic Versioning](https://semver.org/sp
 ## [0.1.8-alpha] - unreleased
 
 ### Added
+- Protect settings from overwrite after load failure
 
 ### Changed